requirements.txt
six==1.10.0
future==0.15.2
requests==2.9.1
boto==2.39.0
gbdx-auth==0.2.0
pygeoif==0.6
pytest==2.9.1
pytest-runner==2.7
sympy==1.0
vcrpy
ndg-httpsclient==0.4.2
bumpversion
docker-py==1.10.4
toposort==1.4